Norminette: split vector
