+++ glmath/Makefile
@@ -2,6 +2,11 @@ NAME=libglmath.a
 
 SRCS=	scene.c \
 		vector.c \
+		vector_norm.c \
+		vector_add_sub.c \
+		vector_angle_op.c \
+		vector_scalar_op.c \
+		vector_product.c \
 		color.c \
 		color2.c \
 		ray.c \
